Add files via upload
The main aspect of this project, “Vision for understanding” intends to identify the facial features to streamline and enhance the identification of autism spectrum disorder (ASD). The neurological and developmental condition known as ASD has an impact on behaviour, learning, and communication. Effective intervention depends on early diagnosis, but this is frequently a drawn-out to many people in UK.